requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg==3.1.18
psycopg_pool==3.2.1
python-dotenv==1.0.1
